pyproject.toml
[tool.poetry]
name = "telegram-bot-example"
version = "0.1.0"
description = ""
authors = ["MDCIrony <miguel.acastillodiaz@outlook.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "20.5"
pydantic = "2.3"
pydantic-settings = "2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
